requirements.txt
pandas>=2.0.0
requests>=2.31.0
apache-airflow>=2.7.0
google-cloud-bigquery>=3.11.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
apache-airflow-providers-google>=10.0.0
